[fix]: register_test.js 삭제 -> register.js 수정
