+++ api/routes/test/check.js
@@ -1,5 +1,4 @@
 import express from 'express'
-import registerTest from '../../../services/auth/register_test'
 
 const router = express.Router()
 
@@ -12,7 +11,5 @@ router.get('/', (req, res) => {
     info: req.decoded,
   })
 })
-
-router.post('/registertest', registerTest)
 
 export default router
